+++ Tracker-iOS/Final Classes/EmojiCell.swift
