+++ Tracker-iOS/Final Classes/StatisticView.swift
@@ -1,0 +1,93 @@
+import UIKit
+
+final class StatisticView: UIView {
+    // MARK: - Layout elements
+    
+    private let numberLabel: UILabel = {
+        let label = UILabel()
+        label.translatesAutoresizingMaskIntoConstraints = false
+        label.font = UIFont.boldSystemFont(ofSize: 34)
+        return label
+    }()
+    
+    private let nameLabel: UILabel = {
+        let label = UILabel()
+        label.translatesAutoresizingMaskIntoConstraints = false
+        label.font = UIFont.systemFont(ofSize: 12, weight: .medium)
+        return label
+    }()
+    
+    // MARK: - Properties
+    private var number: Int {
+        didSet {
+            numberLabel.text = "\(number)"
+        }
+    }
+    private var name: String {
+        didSet {
+            nameLabel.text = name
+        }
+    }
+    
+    // MARK: - Lifecycle
+    required init(number: Int = 0, name: String) {
+        self.number = number
+        self.name = name
+        
+        super.init(frame: .zero)
+        setNumber(number)
+        setName(name)
+        configureViews()
+        configureConstraints()
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+    override func layoutSubviews() {
+        super.layoutSubviews()
+        setupBorder()
+    }
+    
+    // MARK: - Functions
+    func setNumber(_ number: Int) {
+        self.number = number
+    }
+    
+    func setName(_ name: String) {
+        self.name = name
+    }
+}
+
+// MARK: - Layout methods
+extension StatisticView {
+    
+    func configureViews() {
+        translatesAutoresizingMaskIntoConstraints = false
+        [numberLabel, nameLabel].forEach { addSubview($0) }
+    }
+    
+    func setupBorder() {
+        gradientBorder(
+            width: 1,
+            colors: UIColor.gradient,
+            startPoint: .unitCoordinate(.left),
+            endPoint: .unitCoordinate(.right),
+            andRoundCornersWithRadius: 12
+        )
+    }
+    
+    func configureConstraints() {
+        NSLayoutConstraint.activate([
+            numberLabel.leadingAnchor.constraint(equalTo: leadingAnchor, constant: 12),
+            numberLabel.topAnchor.constraint(equalTo: topAnchor, constant: 12),
+            numberLabel.trailingAnchor.constraint(equalTo: trailingAnchor, constant: -12),
+            nameLabel.leadingAnchor.constraint(equalTo: numberLabel.leadingAnchor),
+            nameLabel.topAnchor.constraint(equalTo: numberLabel.bottomAnchor, constant: 8),
+            nameLabel.trailingAnchor.constraint(equalTo: numberLabel.trailingAnchor),
+            nameLabel.bottomAnchor.constraint(equalTo: bottomAnchor, constant: -12),
+        ])
+    }
+}
+
