+++ Tracker-iOS/Helpers/AnalyticsService.swift
@@ -8,9 +8,25 @@ struct AnalyticsService {
         YMMYandexMetrica.activate(with: configuration)
     }
     
-    func report(event: String, params : [AnyHashable : Any]) {
-        YMMYandexMetrica.reportEvent(event, parameters: params, onFailure: { error in
+    func report(event: Events, params : [AnyHashable : Any]) {
+        YMMYandexMetrica.reportEvent(event.rawValue, parameters: params, onFailure: { error in
             print("REPORT ERROR: %@", error.localizedDescription)
         })
     }
+}
+
+//MARK: - Enums for analitics
+
+enum Events: String, CaseIterable {
+    case open = "open"
+    case close = "close"
+    case click = "click"
+}
+
+enum Items: String, CaseIterable {
+    case add_track = "add_track"
+    case track = "track"
+    case filter = "filter"
+    case edit = "edit"
+    case delete = "delete"
 }
