+++ Tracker-iOS/Resource/en-GB.lproj/InfoPlist.strings
@@ -1,0 +1,1 @@
+"CFBundleDisplayName" = "Tracker-iOS";
