+++ Tracker-iOS/Resource/en-GB.lproj/Localizable (Russian).stringsdict
