+++ Tracker-iOS/Resource/ru.lproj/Localizable.strings
@@ -7,6 +7,7 @@
 "TrackersViewController.deleteTracker" = "Уверены что хотите удалить трекер?";
 "TrackersViewController.pin" = "Закрепить";
 "TrackersViewController.unPin" = "Открепить";
+"TrackersViewController.datePicker" = "ru_RU";
 
 // StatisticViewController
 "StatisticViewController.title" = "Статистика";
